refactor(nr-app): move shared map geometry types into map.utils
